moved hive auth actions as into a hook
